// NestJS-Backend-api/src/auth/interfaces/auth-response.interface.ts
import { UserRole } from '@prisma/client';

export interface AuthResponse {
  accessToken: string;
  refreshToken: string;
  user: {
    id: string;
    email: string;
    role: UserRole;
    isSupervisor: boolean;
    profile?: {
      firstName: string;
      lastName: string;
      avatar?: string;
    };
  };
}